+++ tests/amethyst_test/src/fixture/sprite_render_animation_fixture.rs
@@ -4,10 +4,12 @@ use amethyst::{
     },
     assets::{AssetStorage, Handle, Loader},
     ecs::prelude::*,
-    renderer::SpriteRender,
+    renderer::{MaterialTextureSet, Sprite, SpriteRender, SpriteSheet, Texture},
 };
 
 use EffectReturn;
+
+const TEXTURE_ID: u64 = 0;
 
 /// Fixture to test sprite render animation loading.
 #[derive(Debug)]
@@ -22,9 +24,23 @@ impl SpriteRenderAnimationFixture {
     pub fn effect(world: &mut World) {
         // Load the animation.
         let animation_handle = {
+            // This invocation sequence of read_resource / write_resource is to satisfy the borrow
+            // checker, since we don't have NLL yet.
+            let tex_handle = world.read_resource::<Loader>().load_from_data(
+                [0.5; 4].into(),
+                (),
+                &world.read_resource::<AssetStorage<Texture>>(),
+            );
+            world
+                .write_resource::<MaterialTextureSet>()
+                .insert(TEXTURE_ID, tex_handle);
+
+            let loader = world.read_resource::<Loader>();
+            let sprite_sheet_handle =
+                loader.load_from_data(Self::sprite_sheet(), (), &world.read_resource());
             let sprite_sheet_sampler = Sampler {
                 input: vec![0.0],
-                output: vec![SpriteRenderPrimitive::SpriteSheet(0)],
+                output: vec![SpriteRenderPrimitive::SpriteSheet(sprite_sheet_handle)],
                 function: InterpolationFunction::Step,
             };
             let sprite_index_sampler = Sampler {
@@ -33,7 +49,6 @@ impl SpriteRenderAnimationFixture {
                 function: InterpolationFunction::Step,
             };
 
-            let loader = world.read_resource::<Loader>();
             let sprite_sheet_animation_handle =
                 loader.load_from_data(sprite_sheet_sampler, (), &world.read_resource());
             let sprite_index_animation_handle =
@@ -76,5 +91,17 @@ impl SpriteRenderAnimationFixture {
 
         let store = world.read_resource::<AssetStorage<Animation<SpriteRender>>>();
         assert!(store.get(animation_handle).is_some());
+    }
+
+    fn sprite_sheet() -> SpriteSheet {
+        SpriteSheet {
+            texture_id: TEXTURE_ID,
+            sprites: vec![Sprite {
+                width: 10.0,
+                height: 10.0,
+                offsets: [5.; 2],
+                tex_coords: [0.0, 1.0, 0.0, 1.0].into(),
+            }],
+        }
     }
 }