+++ docs/CHANGELOG.md
@@ -17,10 +17,16 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `PrefabData` can now be derived for many situations, see the book for more information ([#1035])
 * Support for DirectionalLight and SpotLight in PBM pass. ([#1074], [#1081])
 * `UiWidget` variant `Custom` for custom composited widgets ([#1112])
+* `amethyst_ui::get_default_font` supports loading system font from Path. ([#1108])
 
 ### Changed
 
+* `Material` animations now directly use `Handle<Texture>` instead of using indirection. ([#1089])
+* `SpriteRenderPrimitive::SpriteSheet` now takes `Handle<SpriteSheet>` instead of a `u64` ID. ([#1089])
+
 ### Removed
+
+* `SpriteSheetSet` is removed as it is no longer needed. ([#1089])
 
 ### Fixed
 
@@ -31,6 +37,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1074]: https://github.com/amethyst/amethyst/pull/1074
 [#1081]: https://github.com/amethyst/amethyst/pull/1081
 [#1112]: https://github.com/amethyst/amethyst/pull/1112
+[#1089]: https://github.com/amethyst/amethyst/pull/1089
+[#1108]: https://github.com/amethyst/amethyst/pull/1108
 
 ## [0.9.0] - 2018-10
 ### Added